fastapi_sqlmodel/pyproject.toml
[tool.poetry]
name = "fastapi-sqlmodel"
version = "0.1.0"
description = ""
authors = ["Muhammad Saad  <saadjameel2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.18"
uvicorn = "^0.30.0"
httpie = "^3.2.2"
requests = "^2.32.3"
pydantic = "^2.7.2"
sqlalchemy = "^2.0.30"
httpx = "^0.27.0"
psycopg2-binary = "^2.9.9"
databases = "^0.9.0"
python-dotenv = "^1.0.1"
docker = "^7.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
